// BassLines.Web/src/data/src/models/CreatePlaylistRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BassLines
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreatePlaylistRequest
 */
export interface CreatePlaylistRequest {
    /**
     * 
     * @type {string}
     * @memberof CreatePlaylistRequest
     */
    id?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreatePlaylistRequest
     */
    trackUris?: Array<string> | null;
    /**
     * 
     * @type {string}
     * @memberof CreatePlaylistRequest
     */
    userId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreatePlaylistRequest
     */
    name?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CreatePlaylistRequest
     */
    _public?: boolean | null;
    /**
     * 
     * @type {boolean}
     * @memberof CreatePlaylistRequest
     */
    collaborative?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof CreatePlaylistRequest
     */
    description?: string | null;
}

export function CreatePlaylistRequestFromJSON(json: any): CreatePlaylistRequest {
    return CreatePlaylistRequestFromJSONTyped(json, false);
}

export function CreatePlaylistRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreatePlaylistRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'trackUris': !exists(json, 'trackUris') ? undefined : json['trackUris'],
        'userId': !exists(json, 'userId') ? undefined : json['userId'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        '_public': !exists(json, 'public') ? undefined : json['public'],
        'collaborative': !exists(json, 'collaborative') ? undefined : json['collaborative'],
        'description': !exists(json, 'description') ? undefined : json['description'],
    };
}

export function CreatePlaylistRequestToJSON(value?: CreatePlaylistRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'trackUris': value.trackUris,
        'userId': value.userId,
        'name': value.name,
        'public': value._public,
        'collaborative': value.collaborative,
        'description': value.description,
    };
}